pyproject.toml
[tool.poetry]
name = "pandora2-0"
version = "2.0"
description = "reverse image search bot"
authors = ["None"]
license = "None"
readme = "README.md"
packages = [{include = "pandora2"}]

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^2.24"
numpy = "^1.24.1"
bs4 = "^0.0.1"
motor = "^3.1.1"
uuid = "^1.30"
Pillow = "^9.4.0"
ImageHash = "^4.3.1"
requests = "^2.28.1"
PixivPy-Async = "^1.2.14"
httpx = "^0.23.3"
pymongo = "^4.3.3"
python-dateutil = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
